third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/svg/text/scripted/transform-dynamic-update-expected.txt

This is a testharness.js-based test.
FAIL Dynamic update of transform CSS property assert_equals: expected 185.796875 but got 185
Harness: the test ran to completion.
